% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rmarkdown.R
\name{render_rmarkdown}
\alias{render_rmarkdown}
\title{Render (executed) checks to an html document with \href{https://rmarkdown.rstudio.com/}{R Markdown}}
\usage{
render_rmarkdown(
  path_checks,
  directory_output,
  path_template = system.file("report-templates/rmarkdown-1/report-1.Rmd", package =
    "trawler")
)
}
\arguments{
\item{path_checks}{The file path of the output of
\code{\link[=execute_checks]{execute_checks()}}.  Required.}

\item{directory_output}{A writable directory to save the report outputs.
The html file is the primary output.  The markdown file is sometimes
useful too.  Required.}

\item{path_template}{The
\href{https://rmarkdown.rstudio.com/articles_intro.html}{rmd file},
which serves as the report template.  Defaults to the
\href{https://github.com/OuhscBbmc/trawler/tree/main/inst/report-templates/rmarkdown-1/report-1.Rmd}{standard template}
built into the trawler package.}
}
\description{
Render (executed) checks to an html document with \href{https://rmarkdown.rstudio.com/}{R Markdown}
}
\examples{
# Step 0: define paths.
#   So this package example executes on every machine, temp files are used.

# Replace the two paths for your specific project
path_data    <- system.file("datasets/pt-event-biochemical.rds", package = "trawler")
path_defs    <- system.file("checks/checks-biochemical.yml", package = "trawler")

# In your real code, this will probably in a static directory.
# If PHI or sensitive info is contained, use a secure location.
directory_output <- file.path(tempdir(check = TRUE), "trawler")
if (!dir.exists(directory_output)) dir.create(directory_output)
path_checks <- tempfile("trawler-checks-", fileext = ".rds", tmpdir = directory_output)

# Step 1: load the check definitions and the dataset to test
ds_pt_event  <- readr::read_rds(path_data)
checks       <- load_checks(path_defs)

# Step 2: execute the checks and save to an rds file
execute_checks(ds_pt_event, checks) |>
  readr::write_rds(path_checks)

# Step 3: render checks as an html report with R Markdown
render_rmarkdown(path_checks, directory_output)

# For the sake of this example, clean up temp files.
#    You probably do NOT want this line in your real code.
unlink(directory_output, recursive = TRUE)

}
