% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-checks.R
\name{execute_checks}
\alias{execute_checks}
\title{Execute the checks against a dataset to return a nested list.}
\usage{
execute_checks(ds, checks, origin)
}
\arguments{
\item{ds}{The \link{data.frame} to be checked.  Required.}

\item{checks}{The \link{list} describing the check.  Is the output of}

\item{origin}{The origin of the dataset.
Currently supports "csv" and "REDCap". Required.
\code{\link[=load_checks]{load_checks()}}.  Required.}
}
\description{
Execute the checks against a dataset to return a nested list.
}
\examples{
# Replace the two paths for your specific project
path_data    <- system.file("datasets/pt-event-biochemical.rds", package = "trawler")
path_checks  <- system.file("checks/checks-biochemical.yml", package = "trawler")

ds_pt_event  <- readr::read_rds(path_data)
checks       <- load_checks(path_checks, origin = "REDCap")

execute_checks(ds_pt_event, checks, origin = "REDCap")

# saveRDS(execute_checks(ds_pt_event, checks), "inst/derived/biochemical.rds")

}
