% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-checks.R
\name{execute_checks}
\alias{execute_checks}
\title{Load and parse checks in a yaml file, and return a list
of useful objects.}
\usage{
execute_checks(ds, checks)
}
\arguments{
\item{ds}{The \link{data.frame} to be checked.  Required.}

\item{checks}{The \link{list} describing the check.  Is the output of
\code{\link[=load_checks]{load_checks()}}.  Required.}
}
\description{
Load and parse checks in a yaml file, and return a list
of useful objects.
}
\examples{
# Replace the paths for your specific project
path_data    <- system.file("datasets/pt-event-biochemical.rds", package = "trawler")
path_checks  <- system.file("checks/checks-biochemical.yml", package = "trawler")

ds_pt_event  <- readr::read_rds(path_data)
checks       <- load_checks(path_checks)

execute_checks(ds_pt_event, checks)

}
