% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-checks.R
\name{load_checks}
\alias{load_checks}
\title{Load and parse checks in a yaml file, and return a list
of useful objects.}
\usage{
load_checks(path_checks)
}
\arguments{
\item{path_checks}{The file path to the yaml file specifying the smells
and rules.}
}
\description{
Load and parse checks in a yaml file, and return a list
of useful objects.
}
