% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawler-package.R
\docType{package}
\name{trawler-package}
\alias{trawler}
\alias{trawler-package}
\title{trawler: Inspect a dataset}
\description{
Drag a dataset to report illegal or suspicious values.
}
\author{
\strong{Maintainer}: Will Beasley \email{wibeasley@hotmail.com} (\href{https://orcid.org/0000-0002-5613-5006}{ORCID})

Authors:
\itemize{
  \item Maleeha Shahid
}

}
\keyword{internal}
